added better test image
also the imgProcessing is writing to the database directly, it won't put the database file in the place where pathfinding is expecting though, it needs to be moved into vehiclePi folder before running pathFinding.

Unless you are erunning visual studios which decides to always make the root folder the one you opened in visual studios which is fairly unhelpful but convinient in this case since it will always use testDB.db from top level folder.